Moved arrays to their own file and updated imports
