+++ cellst/__init__.py
@@ -1,9 +1,9 @@
-__all__ = ['orchestrator', 'pipeline', 'process', 'segment',
+__all__ = ['core.orchestrator', 'core.pipeline', 'process', 'segment',
            'track', 'extract', 'evaluate']
 
 # Import user-facing classes
-from .orchestrator import Orchestrator
-from .pipeline import Pipeline
+from .core.orchestrator import Orchestrator
+from .core.pipeline import Pipeline
 from .process import Processor
 from .segment import Segmenter
 from .track import Tracker
@@ -13,5 +13,5 @@ from .evaluate import Evaluator
 # Import important types
 from .utils import _types as cst_types
 from .utils._types import Image, Mask, Track, Arr, Same
-from .utils._types import Condition, Experiment
+from .core.arrays import Condition, Experiment
 from .utils.slurm_utils import SlurmController
