+++ cellst/core/pipeline.py
@@ -13,7 +13,7 @@ import imageio as iio
 
 from cellst.operation import Operation
 from cellst.utils._types import Image, Mask, Track, Arr, ImageContainer
-from cellst.utils._types import Experiment
+from cellst.core.arrays import Experiment
 from cellst.utils.process_utils import condense_operations, extract_operations
 from cellst.utils.utils import folder_name
 from cellst.utils.log_utils import get_logger, get_console_logger
